requirements.txt
Django==3.1
djangorestframework==3.11.1
gunicorn==20.0.2
whitenoise==6.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
dj-database-url==0.5.0